requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Login==0.6.3
Flask-WTF==1.2.2
python-dotenv==1.2.1
Werkzeug==3.1.3
qrcode[pil]==8.2
Pillow==12.0.0
WTForms==3.2.1
psycopg2-binary==2.9.11
